backend/requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=4.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
sse-starlette>=2.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
beautifulsoup4>=4.12.0
curl_cffi>=0.7.0
groq
